chore: refactor new contract page
