+++ apps/web/src/app/[lang]/app/[type]/contracts/contracts/new-contract/utils.ts
@@ -1,0 +1,6 @@
+import type { UniRefund_ContractService_Contracts_ContractHeaders_ContractHeaderCreateDto as ContractHeaderCreateDto } from "@ayasofyazilim/saas/ContractService";
+
+export type MerchantStepFormDataDto = Pick<
+  ContractHeaderCreateDto,
+  "merchantId" | "addressId"
+>;
